cli/requirements.txt
requests
python-dotenv
crawl4ai
google-generativeai